requirements.txt

boto3
streamlit
python-dotenv
Pillow
